yolo_gui/requirements.txt
PySide6>=6.5.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
PyYAML>=6.0
